Tweak implementation for CMHW vibrator controls/permissions.
This was previously implemented but a permissions issue was causing a Settings FC on boot for the M8.  Needed to do this manually as init.qcom.rc is stored in a different location vs. Shamu where it is at root of device tree.

Original commit from which this was authored:  https://github.com/BrokenROM/device_moto_shamu/commit/71acbbe52027a5b73fce656e4d08e54044aeacaa

~ StayBroken